Extract UrlArg interface
