+++ packages/moxb/src/routing/index.ts
@@ -10,12 +10,19 @@ export * from './StateSpace';
 export {
     UrlArgDefinition,
     UrlArgTypeDef,
+    UrlArg,
+} from './UrlArg';
+
+export {
     URLARG_TYPE_STRING,
     URLARG_TYPE_ORDERED_STRING_ARRAY,
     URLARG_TYPE_UNORDERED_STRING_ARRAY,
     URLARG_TYPE_BOOLEAN,
     URLARG_TYPE_OBJECT,
     URLARG_TYPE_URLENCODED,
+} from './UrlArgTypes';
+
+export {
     /*
     Change as ArgChange,
     getArgValue,
@@ -26,5 +33,6 @@ export {
     createArgSetter, createArgSetterWithSideEffects,
     Change,
     */
-    UrlArg,
-} from './urlArg';
+    UrlArgImpl,
+} from './UrlArgImpl';
+
